dom/cache/test/mochitest/test_chrome_constructor.html

<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
<!DOCTYPE HTML>
<html>
<head>
  <title>Validate Interfaces Exposed to Workers</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">
  var Cu = Components.utils;
  Cu.import('resource://gre/modules/Services.jsm');
  SimpleTest.waitForExplicitFinish();

  SpecialPowers.pushPrefEnv({
    "set": [["dom.caches.enabled", true],
            ["dom.caches.testing.enabled", true]],
  }, function() {
    // attach to a different origin's CacheStorage
    var url = 'http://example.com/';
    var uri = Services.io.newURI(url, null, null);
    var principal = Services.scriptSecurityManager.createCodebasePrincipal(uri, {});
    var storage = new CacheStorage('content', principal);

    // verify we can use the other origin's CacheStorage as normal
    var req = new Request('http://example.com/index.html');
    var res = new Response('hello world');
    var cache;
    storage.open('foo').then(function(c) {
      cache = c;
      ok(cache, 'storage should create cache');
      return cache.put(req, res.clone());
    }).then(function() {
      return cache.match(req);
    }).then(function(foundResponse) {
      return Promise.all([res.text(), foundResponse.text()]);
    }).then(function(results) {
      is(results[0], results[1], 'cache should contain response');
      return storage.delete('foo');
    }).then(function(deleted) {
      ok(deleted, 'storage should delete cache');
      SimpleTest.finish();
    });
  });
</script>
</body>
</html>
